pyproject.toml
[project]
name = "ms-aitour"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "agent-framework==1.0.0rc1",
    "agent-framework-azurefunctions==1.0.0b260219",
    "agent-framework-core==1.0.0rc1",
    "agent-framework-durabletask==1.0.0b260219",
    "aiohttp>=3.13.3",
    "azure-ai-projects[agents]>=2.0.0b3",
    "azure-identity>=1.25.2",
    "azure-search-documents==11.7.0b2",
    "azure-storage-blob>=12.28.0",
    "azure-storage-file-datalake>=12.23.0",
    "google-api-python-client>=2.190.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "jira==3.10.5",
    "msgraph-sdk>=1.55.0",
    "openai>=2.21.0",
    "opentelemetry-semantic-conventions-ai==0.4.13",
    "pandas>=3.0.1",
    "pygithub>=2.8.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
]

[tool.uv]
prerelease = "allow"
